requirements.txt
hydra-core==1.3.2
numpy==1.24.4
omegaconf==2.3.0
sentencepiece==0.1.99
torch==2.0.1
tqdm==4.66.1
transformers==4.32.0
wandb==0.15.8